tests: Split RCT validation tests into new file.
