+++ src/wallet/test/hdwallet_test_fixture.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2020 The Particl Core developers
+// Copyright (c) 2017-2021 The Particl Core developers
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
@@ -12,7 +12,7 @@
 
 class CHDWallet;
 
-/** Testing setup and teardown for wallet.
+/** Testing setup and teardown for particl wallet.
  */
 struct HDWalletTestingSetup: public TestingSetup {
     explicit HDWalletTestingSetup(const std::string& chainName = CBaseChainParams::MAIN);
@@ -23,6 +23,17 @@ struct HDWalletTestingSetup: public TestingSetup {
     std::shared_ptr<CHDWallet> pwalletMain;
     std::unique_ptr<interfaces::Handler> m_chain_notifications_handler;
 };
+
+struct StakeTestingSetup: public HDWalletTestingSetup {
+    StakeTestingSetup(const std::string& chainName = CBaseChainParams::REGTEST):
+        HDWalletTestingSetup(chainName)
+    {
+        SetMockTime(0);
+    }
+};
+
+void StakeNBlocks(CHDWallet *pwallet, size_t nBlocks);
+uint256 AddTxn(CHDWallet *pwallet, CTxDestination &dest, OutputTypes input_type, OutputTypes output_type, CAmount amount, CAmount exploit_amount=0, std::string expect_error="");
 
 #endif // PARTICL_WALLET_TEST_HDWALLET_TEST_FIXTURE_H
