+++ src/wallet/test/stake_tests.cpp
@@ -8,9 +8,6 @@
 
 #include <wallet/test/hdwallet_test_fixture.h>
 #include <chainparams.h>
-#include <miner.h>
-#include <pos/miner.h>
-#include <timedata.h>
 #include <coins.h>
 #include <net.h>
 #include <validation.h>
@@ -27,49 +24,8 @@
 #include <boost/test/unit_test.hpp>
 
 
-struct StakeTestingSetup: public HDWalletTestingSetup {
-    StakeTestingSetup(const std::string& chainName = CBaseChainParams::REGTEST):
-        HDWalletTestingSetup(chainName)
-    {
-        SetMockTime(0);
-    }
-};
-
 BOOST_FIXTURE_TEST_SUITE(stake_tests, StakeTestingSetup)
 
-
-void StakeNBlocks(CHDWallet *pwallet, size_t nBlocks)
-{
-    size_t nStaked = 0;
-    size_t k, nTries = 10000;
-    for (k = 0; k < nTries; ++k) {
-        int nBestHeight = WITH_LOCK(cs_main, return ::ChainActive().Height());
-
-        int64_t nSearchTime = GetAdjustedTime() & ~Params().GetStakeTimestampMask(nBestHeight+1);
-        if (nSearchTime <= pwallet->nLastCoinStakeSearchTime) {
-            std::this_thread::sleep_for(std::chrono::milliseconds(250));
-            continue;
-        }
-
-        std::unique_ptr<CBlockTemplate> pblocktemplate = pwallet->CreateNewBlock();
-        BOOST_REQUIRE(pblocktemplate.get());
-
-        if (pwallet->SignBlock(pblocktemplate.get(), nBestHeight+1, nSearchTime)) {
-            CBlock *pblock = &pblocktemplate->block;
-
-            if (CheckStake(pblock)) {
-                nStaked++;
-            }
-        }
-
-        if (nStaked >= nBlocks) {
-            break;
-        }
-        std::this_thread::sleep_for(std::chrono::milliseconds(250));
-    }
-    BOOST_REQUIRE(k < nTries);
-    SyncWithValidationInterfaceQueue();
-}
 
 static void AddTxn(CHDWallet *pwallet, CTxDestination &dest, OutputTypes output_type, CAmount amount)
 {
